refactor(part): 파트 리비전 라이프사이클 엔티티/DTO/서비스 정합성 강화
- 파트 상세 응답, 쿼리, 서비스, 유스케이스를 리비전 중심으로 정비해 초안과 상태 반영 경로를 일관화

- PartRevision 도메인 및 파트 생성 흐름 업데이트로 리비전 전환 로직을 명확히 함

- 테스트 및 합성 실행 경로도 변경된 리비전 정책에 맞춰 동기화

+++ src/main/java/com/fabbitinc/server/application/mapping/support/MappingGenerationSupport.java
@@ -255,7 +255,7 @@ public class MappingGenerationSupport {
         String quantity = findHeader(headers, List.of("수량", "qty", "quantity"));
         if (quantity != null) {
             relColumns.put("quantity", quantity);
-            relColumnTypes.put("quantity", PropertyDataType.INTEGER);
+            relColumnTypes.put("quantity", PropertyDataType.FLOAT);
         }
 
         return new RelationMappingDto(
