+++ src/main/java/com/fabbitinc/server/application/part/query/result/BomTreeResult.java
@@ -2,6 +2,7 @@ package com.fabbitinc.server.application.part.query.result;
 
 import com.fabbitinc.server.application.part.model.BomDirection;
 import com.fabbitinc.server.domain.part.model.PartLifecycleState;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -19,7 +20,7 @@ public record BomTreeResult(
             String unit,
             String category,
             PartLifecycleState lifecycleState,
-            int quantity,
+            BigDecimal quantity,
             List<Node> children
     ) {
     }
